+++ src/idi_corporate_structure/extractor.py
@@ -9,11 +9,11 @@ from abc import ABC, abstractmethod
 
 # Third-party imports
 from bs4 import BeautifulSoup, Comment
+from idi_ftm2j_shared.logs import get_logger
 
 # Application imports
-from idi_corporate_structure.common.api import OpenAiClient
-from idi_corporate_structure.common.logs import get_logger
-from idi_corporate_structure.processor.types import Filing, Subsidiary
+from idi_corporate_structure.api import OpenAiClient
+from idi_corporate_structure.types import Filing, Subsidiary
 
 _PROMPTS = importlib.resources.files("idi_corporate_structure.processor.prompts")
