+++ src/idi_corporate_structure/prompts/gpt_extractor_system.txt
