+++ src/idi_corporate_structure/types.py
